pyproject.toml
[project]
name = "anime-recommender"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.3.5",
    "langchain>=1.0.7",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.0.1",
    "langchain-huggingface>=1.0.1",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.51.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.5",
]
